requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
anthropic==0.43.0
httpx==0.28.1
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
deepgram-sdk==3.9.0
